\name{undocumented-methods}
\docType{methods}
%\alias{ccoef,ANY-method}
%\alias{ccoef-method}
%
%
% 201607, 201608:
%\alias{psint,pvgam-method}
%\alias{summary,pvgam-method}
%\alias{show,summary.pvgam-method}
%\alias{df.residual,pvgam-method}
\alias{typeTS,vglm-method}
%
%
% This does not work (need one line for each one):
  %\alias{trplot,qrrvglm,uqo-method}
%
%
%
\title{ Undocumented Methods Functions in \pkg{VGAMextra} }

\description{
  Undocumented methods functions are aliased here.
  In the \pkg{VGAMextra} package there are currently some
  objects/methods/classes which are currently internal and/or
  undocumented. The help file suppresses the warnings when the package is
  'CHECK'ed.
  
}
%\usage{
  %  \S4method{ccoef}{rrvgam,Coef.rrvgam,rrvglm,qrrvglm,
    %                   Coef.rrvglm,Coef.qrrvglm}(object, ...)
  %}

\section{Methods}{
  Methods/classes will be documented over time.
  
  \describe{
    
    \item{object}{
      This argument is often used, and it is the primary object from which the
      function operates on.
      
    }
  }
}
\keyword{methods}
\keyword{classes}
%\keyword{ ~~ other possible keyword(s)}
\keyword{models}
\keyword{regression}
\keyword{internal}
